Add Sweet alert

+++ GameZone/wwwroot/js/game-index.js
@@ -2,15 +2,48 @@
     $('.js-delete').on('click', function () {
         var btn = $(this);
 
-        $.ajax({
-            url: `/Games/Delete/${btn.data('id')}`,
-            method: 'DELETE',
-            success: function () {
-                alert('success');
+        const swal = Swal.mixin({
+            customClass: {
+                confirmButton: "btn btn-danger mx-2",
+                cancelButton: "btn btn-light"
             },
-            error: function () {
-                alert('error');
-            }
+            buttonsStyling: false
         });
+
+        swal.fire({
+            title: "Are you sure that you want to delete this game?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "Yes, delete it!",
+            cancelButtonText: "No, cancel!",
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                $.ajax({
+                    url: `/Games/Delete/${btn.data('id')}`,
+                    method: 'DELETE',
+                    success: function () {
+                        swal.fire({
+                            title: "Deleted!",
+                            text: "Game has been deleted.",
+                            icon: "success"
+                        })
+                        btn.parents('tr').fadeOut();
+                    },
+                    error: function () {
+                        swal.fire({
+                            title: "Ooooops...!",
+                            text: "Something went wrong.",
+                            icon: "error"
+                        })
+                    }
+                });
+
+              
+            } 
+        });
+       
     });
 });
